testing-app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.9; python_version >= "3.12"
psycopg2-binary==2.9.9; python_version < "3.12"
PyYAML==6.0.2
httpx==0.27.0
celery==5.4.0
redis==5.1.1
alembic==1.13.3
pytest==8.4.1
mypy==1.17.1
ruff==0.12.8
black==25.1.0

